+++ bee-block/src/output/native_token.rs
@@ -40,7 +40,7 @@ impl NativeToken {
 
     /// Returns the amount of the [`NativeToken`].
     #[inline(always)]
-    pub fn amount(&self) -> &U256 {
+    pub fn amount(&self) -> U256 {
         &self.amount
     }
 }